publish and integrate with gem
