+++ gems/ecko-plugins/lib/ecko/plugins/version.rb
@@ -1,0 +1,5 @@
+module Ecko
+  module Plugins
+    VERSION = "0.1.0"
+  end
+end
