+++ gems/ecko-plugins/spec/ecko/plugins_spec.rb
@@ -1,0 +1,9 @@
+RSpec.describe Ecko::Plugins do
+  it "has a version number" do
+    expect(Ecko::Plugins::VERSION).not_to be nil
+  end
+
+  it "does something useful" do
+    expect(false).to eq(true)
+  end
+end
