+++ gems/ecko-plugins/spec/spec_helper.rb
@@ -1,0 +1,14 @@
+require "bundler/setup"
+require "ecko/plugins"
+
+RSpec.configure do |config|
+  # Enable flags like --only-failures and --next-failure
+  config.example_status_persistence_file_path = ".rspec_status"
+
+  # Disable RSpec exposing methods globally on `Module` and `main`
+  config.disable_monkey_patching!
+
+  config.expect_with :rspec do |c|
+    c.syntax = :expect
+  end
+end
